Require tests to explicitly include all fact values
This mostly impacts anywhere the operation expects the fact to return
`None`, which was the default before this change. Now requires each test
to explicitly define the fact return value or an exception will be raised.

As can be seen in the mass update of test files there were a number of
invalid test facts defined that relied on this bahavior, moving to an
explicit definition is a significant improvement.